Moved analyser logic to plans, flattened tests
