+++ .gitignore
@@ -78,4 +78,3 @@ lockfiles/
 
 # custom CI setup 
 report.json
-Pipfile
